backend/requirements.txt
celery
Flask
Flask-SocketIO
redis
gevent
flask-cors
nose
uwsgi

paho-mqtt
